Changed ID handling from int to long, as it was hitting the max int value due to merge checkout/checkin activities for large number of records.

+++ Timelapse/Recognition/Recognizer.cs
@@ -156,7 +156,7 @@ namespace Timelapse.Recognition
     public class detection
     {
         #region Public Properties
-        public int detectionID { get; set; }
+        public long detectionID { get; set; }
         public string category { get; set; } = string.Empty;
         public float conf { get; set; } = 0;
